Release 1.3. Clean stylistically mistake. New parse method with cullture settings. Replace douuble to decimal.

+++ ConfigData.cs
@@ -10,9 +10,10 @@ internal static class ConfigData
     public static string CellAddress { get; private set; } = "C18";
     public static DateTime LastProcessedDate { get; private set; } = new DateTime(2023, 01, 01, 00, 00, 00);
     public static TimeOnly ProcessedTime { get; private set; } = new TimeOnly(05, 00);
-    public static double LastCellValue { get; set; } = -1;
+    public static decimal LastCellValue { get; set; } = -1;
 
-    public static void ReadAppConfig() 
+
+    public static void ReadAppConfig()
     {
         var allConfiguration = ConfigurationManager.AppSettings;
         
@@ -36,7 +37,7 @@ internal static class ConfigData
                     ProcessedTime = TimeOnly.Parse(allConfiguration[keys]); break;
 
                 case ConfigKeys.lastCellValue:
-                    LastCellValue = Double.Parse(allConfiguration[keys]); break;
+                    LastCellValue = Decimal.Parse(allConfiguration[keys]); break;
             }
         }
     }
