+++ DataInfo.cs
@@ -4,11 +4,11 @@ using DayEfficiency;
 
 internal class DataInfo
 {
-    public double LastMonthEfficiency { get; set; }
-    public double CurrentEfficiency { get; set; }
-    public double CurrentMonthEfficiency { get; set; }
+    public decimal LastMonthEfficiency { get; set; }
+    public decimal CurrentEfficiency { get; set; }
+    public decimal CurrentMonthEfficiency { get; set; }
     public DateTime LastUpdateDate { get; set; }
-    public DateTime CurentDate { get; set; }        
+    public DateTime CurentDate { get; set; }
     public TimeOnly ProcessedTime { get; set; }
     public DateTime LastWriteDateSourceFile { get; set; }
 }
